+++ NINA/ViewModel/DockManagerVM.cs
@@ -45,6 +45,7 @@ using NINA.WPF.Base.ViewModel;
 using NINA.Plugin.Interfaces;
 using System.Diagnostics;
 using System.Threading;
+using NINA.Core.Utility.Notification;
 
 namespace NINA.ViewModel {
 
@@ -74,6 +75,8 @@ namespace NINA.ViewModel {
                              IPluginLoader pluginProvider) : base(profileService) {
             LoadAvalonDockLayoutCommand = new AsyncCommand<bool>((object o) => Task.Run(() => InitializeAvalonDockLayout(o)));
             ResetDockLayoutCommand = new RelayCommand(ResetDockLayout, (object o) => _dockmanager != null);
+            BackupDockLayoutCommand = new RelayCommand(BackupDockLayout, (object o) => _dockmanager != null);
+            RestoreDockLayoutFromFileCommand = new RelayCommand(RestoreDockLayoutFromFile);
 
             var initAnchorables = new List<IDockableVM>();
             var initAnchorableInfoPanels = new List<IDockableVM>();
@@ -145,6 +148,50 @@ namespace NINA.ViewModel {
             });
         }
 
+        private void RestoreDockLayoutFromFile(object obj) {
+            try {
+                var dialog = OptionsVM.GetFilteredFileDialog("", "DockBackup.dock.config", "Dock Config|*.dock.config");
+                if (dialog.ShowDialog() == true) {
+                    if(File.Exists(dialog.FileName)) {
+                        lock(lockObj) {
+                            _dockloaded = false;
+                            File.Copy(dialog.FileName, GetDockConfigPath(profileService.ActiveProfile.Id), true);
+                            Notification.ShowInformation(Loc.Instance["LblDockLayoutRestored"]);
+                        }
+                    }
+                }
+            } catch (Exception ex) {
+                Logger.Error(ex);
+                Notification.ShowError(Loc.Instance["LblRestoreDockLayoutFromFileFailed"]);
+            }
+        }
+
+        private void BackupDockLayout(object obj) {
+            try {
+                Microsoft.Win32.SaveFileDialog dialog = new Microsoft.Win32.SaveFileDialog();
+                dialog.InitialDirectory = "";
+                dialog.FileName = "DockBackup.dock.config";
+                dialog.Title = Loc.Instance["LblBackupDockLayout"];
+                dialog.DefaultExt = ".dock.config";
+                dialog.Filter = "Dock Config|*.dock.config";
+                dialog.OverwritePrompt = true;
+
+                if (dialog.ShowDialog().Value) {
+                    if (Directory.Exists(Path.GetDirectoryName(dialog.FileName))) {
+                        lock (lockObj) {
+                            var serializer = new AvalonDock.Layout.Serialization.XmlLayoutSerializer(_dockmanager);
+                            serializer.Serialize(dialog.FileName);
+                            Notification.ShowInformation(Loc.Instance["LblBackupDockLayoutSuccessful"]);
+                        }
+                    }
+                    
+                }
+            } catch(Exception ex) {
+                Logger.Error(ex);
+                Notification.ShowError(Loc.Instance["LblBackupDockLayoutFailed"]);
+            }
+        }
+
         private bool initialized;
 
         public bool Initialized {
@@ -186,6 +233,7 @@ namespace NINA.ViewModel {
 
                     LoadDefaultLayout(serializer);
                     SaveAvalonDockLayout();
+                    Notification.ShowInformation(Loc.Instance["LblDockLayoutReset"]);
                 }
             }
         }
@@ -388,7 +436,8 @@ namespace NINA.ViewModel {
 
         public IAsyncCommand LoadAvalonDockLayoutCommand { get; private set; }
         public ICommand ResetDockLayoutCommand { get; }
-
         public ICommand ClosingCommand { get; private set; }
+        public ICommand BackupDockLayoutCommand { get; private set; }
+        public ICommand RestoreDockLayoutFromFileCommand { get; private set; }        
     }
 }